app/pyproject.toml
[project]
name = "StagyBee"
version = "2.0.0-alpha01"
description = "StagyBee Auto login for JWConf Stage User"
requires-python = ">=3.14"
dependencies = [
    "aiohttp==3.13.2",
    "asgiref==3.11.0",
    "beautifulsoup4==4.14.3",
    "channels-redis[cryptography]==4.3.0",
    "channels[daphne]==4.3.2",
    "django==6.0",
    "django-debug-toolbar==6.1.0",
    "django-environ==0.12.0",
    "django-guardian==3.2.0",
    "django-ninja==1.5.1",
    "django-qr-code==4.2.0",
    "django-widget-tweaks==1.5.0",
    "pytest-asyncio==1.3.0",
    "python-dateutil==2.9.0.post0",
    "redis==7.1.0",
    "tenacity==9.1.2",
    "wheel==0.46.1",
]

[project.optional-dependencies]
postgres = [
    "psycopg>=3.3.2",
]

[dependency-groups]
dev = [
    "django-debug-toolbar>=6.1.0",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "pytest-django>=4.11.1",
    "python-dateutil==2.9.0.post0",
]
